azure/requirements.txt
# rest framework
Django==1.9.6
djangorestframework==3.3.3

# redis cache
redis==2.10.5

# for access redis cache
redisco==0.1.4
django-redis-cache==0.13.1

# for call rest api
httplib2==0.9.2

# for call openstack api
# openstacksdk==0.9.15
# os-client-config==1.29.0
# python-cinderclient==3.5.0

# for unit test
django-nose>=1.4.0
coverage==4.2
mock==2.0.0
unittest_xml_reporting==1.12.0

# for onap logging
onappylog>=1.0.6

# for event
oslo_messaging

# for pecan framework
uwsgi
pecan>=1.2.1
oslo.concurrency>=3.21.0
oslo.config>=4.11.0
oslo.service>=1.25.0
eventlet>=0.20.0
PyYAML>=3.1.0

#azure
azure-mgmt-resource==2.0.0
azure-mgmt-compute==4.0.1
azure-mgmt-authorization==0.50.0
azure-common==1.1.14
